Update index.txt
As for Ubuntu 16.04 Xenial Xerus, the pip installation of virtualenvwrapper is not enough.  For it to work as is described afterwards (#. Set up a virtual environment) it must be installed system wide with apt or else it returns the error "/usr/bin/python: No module named virtualenvwrapper/".  I don't know if this error occurs in other distros.